// WEB/web-app/src/models/DTO_MetricaKPI.ts
export class DTO_MetricaKPI {
    iD_Negocio?: number = 0;
    tituloRegular: string = "";
    tituloNegrita: string = "";
    ordenTitulos: string = "";
    txtColor: string = "";
    bgColor: string = "";
    valorRegular: string = "";
    valorNegrita: string = "";
    ordenValores: string = "";
    icono: string = "";
    info: string = "";
    filtro: string = "";
}